Use wrapper type for `expiresIn` field of `PollEntity`
This fixes the missing `expires_in` key in the poll object of the
`/statuses` POST response causing `expiresIn` in the Java
representation of that response to be set to 0 instead of `null`.

The value `null` correctly causes `copyFields()` to skip the `expiresIn`
field when copying over values from the response to be saved in the
database. The integer 0 on the other hand would be copied as-is and thus
stored in the database, even though this is not the value the user
originally set before posting the status.